Don't repeat code in onOptionSelect hook
The plugin's definition of onOptionSelect only needs to provide the new
behavior for handling disabled options, and should fall through to the
underlying implementation for the create/select choices rather than
copy-and-pasting that code into the plugin.

This is important so that when fixes are made in Selectize.js they don't
need to be replicated here.