Allow disabled options to be changed at runtime
The initial version of this plugin only allowed disabled options to be
set when a Selectize instance was created, but in some circumstances you
may need to disable choices on the fly.

This change adds four new methods to the API to allow this:
- disableOptions(values): Add the provided values to the disabled list.
- enableOptions(values): Remove the provided values to the disabled list.
- setDisabledOptions(values): Make the provided values the only ones disabled.
- disabledOptions(): Retrieve an array of the disabled options.

This is similar to @ghost/734537b6947a1b85ac07ec934a749b4d4d43fa35 but
just contains new methods without overwriting any of the original module.

Fixes mondorobot/selectize-disable-options#2.